Synced energy and energy gain rate

+++ Assets/Scripts/GameScripts/CardEnergySource.cs
@@ -6,8 +6,7 @@ using UnityEngine;
 public class CardEnergySource : MonoBehaviour
 {
     public string nameOfObjectToSpawn;
-
-    public float generationRate;
+    public int generationRate;
     public float lifetime;
     public int cost;
